+++ lib/flapjack/data/contact.rb
@@ -246,7 +246,9 @@ module Flapjack
         if rules.all? {|r| r.is_specific? } # also true if empty
           rule = self.add_notification_rule({
               :entities           => [],
+              :regex_entities     => [],
               :tags               => Flapjack::Data::TagSet.new([]),
+              :regex_tags         => Flapjack::Data::TagSet.new([]),
               :time_restrictions  => [],
               :warning_media      => ALL_MEDIA,
               :critical_media     => ALL_MEDIA,