+++ lib/flapjack/data/notification.rb
@@ -52,6 +52,7 @@ module Flapjack
 
         tag_data = event.tags.is_a?(Set) ? event.tags.to_a : nil
         notif = {'event_id'       => event.id,
+                 'event_hash'     => event.id_hash,
                  'state'          => event.state,
                  'summary'        => event.summary,
                  'details'        => event.details,
@@ -105,6 +106,7 @@ module Flapjack
 
       def contents
         @contents ||= {'event_id'          => @event_id,
+                       'event_hash'        => @event_hash,
                        'state'             => @state,
                        'summary'           => @summary,
                        'duration'          => @duration,
@@ -141,14 +143,19 @@ module Flapjack
             # matchers are rules of the contact that have matched the current event
             # for time, entity and tags
             matchers = rules.select do |rule|
-              logger.debug("considering rule with entities: #{rule.entities} and tags: #{rule.tags.to_json}")
-              rule_has_tags     = rule.tags     ? (rule.tags.length > 0)     : false
-              rule_has_entities = rule.entities ? (rule.entities.length > 0) : false
+              logger.debug("considering rule with entities: #{rule.entities}, entities regex: #{rule.regex_entities},
+                           tags: #{rule.tags.to_json} and regex tags: #{rule.regex_tags.to_json}")
+              rule_has_tags           = rule.tags           ? (rule.tags.length > 0)           : false
+              rule_has_regex_tags     = rule.regex_tags     ? (rule.regex_tags.length > 0)     : false
+              rule_has_entities       = rule.entities       ? (rule.entities.length > 0)       : false
+              rule_has_regex_entities = rule.regex_entities ? (rule.regex_entities.length > 0) : false
 
-              matches_tags   = rule_has_tags     ? rule.match_tags?(@tags)       : true
-              matches_entity = rule_has_entities ? rule.match_entity?(@event_id) : true
+              matches_tags           = rule_has_tags           ? rule.match_tags?(@tags)               : true
+              matches_regex_tags     = rule_has_regex_tags     ? rule.match_regex_tags?(@tags)         : true
+              matches_entity         = rule_has_entities       ? rule.match_entity?(@event_id)         : true
+              matches_regex_entities = rule_has_regex_entities ? rule.match_regex_entities?(@event_id) : true
 
-              ((matches_entity && matches_tags) || ! rule.is_specific?) &&
+              ((matches_entity && matches_regex_entities && matches_tags && matches_regex_tags) || ! rule.is_specific?) &&
                 rule_occurring_now?(rule, :contact => contact, :default_timezone => default_timezone)
             end