+++ lib/flapjack/data/notification_rule.rb
@@ -12,8 +12,8 @@ module Flapjack
 
       extend Flapjack::Utility
 
-      attr_accessor :id, :contact_id, :entities, :tags, :time_restrictions,
-        :unknown_media, :warning_media, :critical_media,
+      attr_accessor :id, :contact_id, :entities, :regex_entities, :tags, :regex_tags,
+        :time_restrictions, :unknown_media, :warning_media, :critical_media,
         :unknown_blackhole, :warning_blackhole, :critical_blackhole
 
       def self.exists_with_id?(rule_id, options = {})
@@ -74,7 +74,7 @@ module Flapjack
       end
 
       def to_json(*args)
-        self.class.hashify(:id, :contact_id, :tags, :entities,
+        self.class.hashify(:id, :contact_id, :tags, :regex_tags, :entities, :regex_entities,
             :time_restrictions, :unknown_media, :warning_media, :critical_media,
             :unknown_blackhole, :warning_blackhole, :critical_blackhole) {|k|
           [k, self.send(k)]
@@ -87,10 +87,33 @@ module Flapjack
         @entities.include?(event_id.split(':').first)
       end
 
+      # entity names match regex?
+      def match_regex_entities?(event_id)
+        return false unless @regex_entities && @regex_entities.length > 0
+        entity = event_id.split(':').first
+        matches = 0
+        @regex_entities.each do |regex_entity|
+          matches += 1 if /#{regex_entity}/ === entity
+        end
+        matches >= @regex_entities.length
+      end
+
       # tags match?
       def match_tags?(event_tags)
         return false unless @tags && @tags.length > 0
         @tags.subset?(event_tags)
+      end
+
+      # regex_tags match?
+      def match_regex_tags?(event_tags)
+        return false unless @regex_tags && @regex_tags.length > 0
+        matches = 0
+        event_tags.each do |event_tag|
+          @regex_tags.each do |regex_tag|
+            matches += 1 if /#{regex_tag}/ === event_tag
+          end
+        end
+        matches >= @regex_tags.length
       end
 
       def blackhole?(severity)
@@ -112,7 +135,9 @@ module Flapjack
 
       def is_specific?
         (!@entities.nil? && !@entities.empty?) ||
-          (!@tags.nil? && !@tags.empty?)
+          (!@regex_entities.nil? && !@regex_entities.empty?) ||
+          (!@tags.nil? && !@tags.empty?) ||
+          (!@regex_tags.nil? && !@regex_tags.empty?)
       end
 
     private
@@ -137,6 +162,9 @@ module Flapjack
         if rule_data[:tags].is_a?(Array)
           rule_data[:tags] = Flapjack::Data::TagSet.new(rule_data[:tags])
         end
+        if rule_data[:regex_tags].is_a?(Array)
+          rule_data[:regex_tags] = Flapjack::Data::TagSet.new(rule_data[:regex_tags])
+        end
         rule_data
       end
 
@@ -150,12 +178,15 @@ module Flapjack
         return errors unless errors.nil? || errors.empty?
 
         # whitelisting fields, rather than passing through submitted data directly
-        tag_data = rule_data[:tags].is_a?(Set) ? rule_data[:tags].to_a : nil
+        tag_data       = rule_data[:tags].is_a?(Set) ? rule_data[:tags].to_a : nil
+        regex_tag_data = rule_data[:regex_tags].is_a?(Set) ? rule_data[:regex_tags].to_a : nil
         json_rule_data = {
           :id                 => rule_data[:id].to_s,
           :contact_id         => rule_data[:contact_id].to_s,
           :entities           => Oj.dump(rule_data[:entities]),
+          :regex_entities     => Oj.dump(rule_data[:regex_entities]),
           :tags               => Oj.dump(tag_data),
+          :regex_tags         => Oj.dump(regex_tag_data),
           :time_restrictions  => Oj.dump(rule_data[:time_restrictions], :mode => :compat),
           :unknown_media      => Oj.dump(rule_data[:unknown_media]),
           :warning_media      => Oj.dump(rule_data[:warning_media]),
@@ -244,11 +275,23 @@ module Flapjack
                  d[:entities].all? {|e| e.is_a?(String)} ) } =>
         "entities must be a list of strings",
 
+        proc {|d| !d.has_key?(:regex_entities) ||
+               ( d[:regex_entities].nil? ||
+                 d[:regex_entities].is_a?(Array) &&
+                 d[:regex_entities].all? {|e| e.is_a?(String)} ) } =>
+        "regex_entities must be a list of strings",
+
         proc {|d| !d.has_key?(:tags) ||
                ( d[:tags].nil? ||
                  d[:tags].is_a?(Flapjack::Data::TagSet) &&
                  d[:tags].all? {|et| et.is_a?(String)} ) } =>
         "tags must be a tag_set of strings",
+
+        proc {|d| !d.has_key?(:regex_tags) ||
+               ( d[:regex_tags].nil? ||
+                 d[:regex_tags].is_a?(Flapjack::Data::TagSet) &&
+                 d[:regex_tags].all? {|et| et.is_a?(String)} ) } =>
+        "regex_tags must be a tag_set of strings",
 
         proc {|d| !d.has_key?(:time_restrictions) ||
                ( d[:time_restrictions].nil? ||
@@ -318,7 +361,10 @@ module Flapjack
         @contact_id         = rule_data['contact_id']
         tags                = Oj.load(rule_data['tags'] || '')
         @tags               = tags ? Flapjack::Data::TagSet.new(tags) : nil
+        regex_tags          = Oj.load(rule_data['regex_tags'] || '')
+        @regex_tags         = regex_tags ? Flapjack::Data::TagSet.new(regex_tags) : nil
         @entities           = Oj.load(rule_data['entities'] || '')
+        @regex_entities     = Oj.load(rule_data['regex_entities'] || '')
         @time_restrictions  = Oj.load(rule_data['time_restrictions'] || '')
         @unknown_media      = Oj.load(rule_data['unknown_media'] || '')
         @warning_media      = Oj.load(rule_data['warning_media'] || '')