+++ lib/flapjack/notifier.rb
@@ -30,10 +30,8 @@ module Flapjack
 
       @notifications_queue = @config['queue'] || 'notifications'
 
-      @queues = {:email     => @config['email_queue'],
-                 :sms       => @config['sms_queue'],
-                 :jabber    => @config['jabber_queue'],
-                 :pagerduty => @config['pagerduty_queue']}
+      queue_configs = @config.find_all {|k, v| k =~ /_queue$/ }
+      @queues = Hash[queue_configs.map {|k, v| [k[/^(.*)_queue$/, 1], v] }]
 
       notify_logfile  = @config['notification_log_file'] || 'log/notify.log'
       if not File.directory?(File.dirname(notify_logfile))
@@ -133,7 +131,7 @@ module Flapjack
         @notifylog.info("#{event_id} | " +
           "#{notification.type} | #{message.contact.id} | #{media_type} | #{address}")
 
-        unless @queues[media_type.to_sym]
+        unless @queues[media_type.to_s]
           @logger.error("no queue for media type: #{media_type}")
           return
         end
@@ -160,16 +158,15 @@ module Flapjack
         contents_tags = contents['tags']
         contents['tags'] = contents_tags.is_a?(Set) ? contents_tags.to_a : contents_tags
 
-        # TODO consider changing Resque jobs to use raw blpop like the others
         case media_type.to_sym
         when :sms
-          Resque.enqueue_to(@queues[:sms], Flapjack::Gateways::SmsMessagenet, contents)
+          # FIXME(@auxesis): change Resque jobs to use raw blpop
+          Resque.enqueue_to(@queues['sms'], Flapjack::Gateways::SmsMessagenet, contents)
         when :email
-          Resque.enqueue_to(@queues[:email], Flapjack::Gateways::Email, contents)
-        when :jabber
-          @redis.rpush(@queues[:jabber], Oj.dump(contents))
-        when :pagerduty
-          @redis.rpush(@queues[:pagerduty], Oj.dump(contents))
+          # FIXME(@auxesis): change Resque jobs to use raw blpop
+          Resque.enqueue_to(@queues['email'], Flapjack::Gateways::Email, contents)
+        else
+          @redis.rpush(@queues[media_type.to_s], Oj.dump(contents))
         end
       end
     end