+++ spec/lib/flapjack/data/global_spec.rb
@@ -1,8 +1,0 @@
-require 'spec_helper'
-require 'flapjack/data/global'
-
-describe Flapjack::Data::Global do
-
-  it "returns unacknowledged failing checks"
-
-end