+++ spec/lib/flapjack/gateways/jabber_spec.rb
@@ -52,13 +52,14 @@ describe Flapjack::Gateways::Jabber, :logger => true do
   end
 
   it "receives an acknowledgement message" do
-    expect(stanza).to receive(:body).and_return('flapjack: ACKID 876 fixing now duration: 90m')
+    expect(stanza).to receive(:body).and_return('flapjack: ACKID 1f8ac10f fixing now duration: 90m')
     from = double('from')
+    expect(from).to receive(:resource).and_return('sender')
     expect(from).to receive(:stripped).and_return('sender')
     expect(stanza).to receive(:from).and_return(from)
 
     redis = double('redis')
-    expect(redis).to receive(:hget).with('unacknowledged_failures', '876').
+    expect(redis).to receive(:hget).with('check_hashes', '1f8ac10f').
       and_return('main-example.com:ping')
 
     entity_check = double(Flapjack::Data::EntityCheck)
@@ -66,7 +67,7 @@ describe Flapjack::Gateways::Jabber, :logger => true do
 
     expect(Flapjack::Data::Event).to receive(:create_acknowledgement).
       with('main-example.com', 'ping', :summary => 'fixing now',
-           :acknowledgement_id => '876',
+           :acknowledgement_id => '1f8ac10f',
            :duration => (90 * 60), :redis => redis)
 
     expect(Flapjack::Data::EntityCheck).to receive(:for_event_id).
@@ -89,6 +90,7 @@ describe Flapjack::Gateways::Jabber, :logger => true do
                  '<a href="http://example.org/">example.org</a></span>')
 
     from = double('from')
+    expect(from).to receive(:resource).and_return('sender')
     expect(from).to receive(:stripped).and_return('sender')
     expect(stanza).to receive(:from).and_return(from)
 
@@ -121,6 +123,7 @@ describe Flapjack::Gateways::Jabber, :logger => true do
       and_return("flapjack: tell me about \nexample.com")
 
     from = double('from')
+    expect(from).to receive(:resource).and_return('sender')
     expect(from).to receive(:stripped).and_return('sender')
     expect(stanza).to receive(:from).and_return(from)
 
@@ -151,6 +154,7 @@ describe Flapjack::Gateways::Jabber, :logger => true do
   it "receives a message it doesn't understand" do
     expect(stanza).to receive(:body).once.and_return('flapjack: hello!')
     from = double('from')
+    expect(from).to receive(:resource).and_return('sender')
     expect(from).to receive(:stripped).and_return('sender')
     expect(stanza).to receive(:from).and_return(from)